fix(canvas): Improve anchor drag and path hit area
Increase the invisible hit area for paths and anchor points to make them easier to interact with. This includes increasing the `strokeWidth` of the invisible path and the `r` value of the invisible anchor circle.

Additionally, prevent canvas pan events from interfering with anchor dragging by calling `e.stopPropagation()` and `e.preventDefault()` on the `onMouseDown` event of the anchor point's `g` element. This ensures a smoother and more intuitive user experience when manipulating anchors.